requirements.txt
chex
einops
flax
torchaudio
torchvision
torchtext
tensorflow-datasets
pydub
datasets
tqdm
wandb
plotly